My best project up to now.
Everything is explained in the comment before the code on the masterpiece version. The algorithm employed in this problem is extremely beautiful and very efficient. the worst case it might face is when the stock steadily loses value, otherwise it terminates sooner. compared to the other, more straightforward, but silly algorithm I am uploading, it has monstrous efficiency on an array of 100000 random data between -500 and 500. i came up with a way to improve it further through compression: it will greatly improve efficiency for data that is not too volatile. i will now implement it.